+++ app/src/main/java/com/sourcepoint/test_project/MainActivity.java
@@ -1,5 +1,7 @@
 package com.sourcepoint.test_project;
 
+import android.content.SharedPreferences;
+import android.preference.PreferenceManager;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
@@ -24,7 +26,6 @@ public class MainActivity extends AppCompatActivity {
             view.requestLayout();
             mainViewGroup.addView(view);
         }
-
     }
     private void removeView(View view) {
         if(view.getParent() != null)
@@ -33,9 +34,8 @@ public class MainActivity extends AppCompatActivity {
 
 
     private GDPRConsentLib buildGDPRConsentLib() {
-        return GDPRConsentLib.newBuilder(22, "mobile.demo", 2372,"5c0e81b7d74b3c30c6852301",this)
+        return GDPRConsentLib.newBuilder(22, "a-demo-property", 7055,"5c0e81b7d74b3c30c6852301",this)
                 .setStagingCampaign(false)
-                .setTargetingParam("native", "true")
                 .setOnConsentUIReady(view -> {
                     showView(view);
                     Log.i(TAG, "onConsentUIReady");
@@ -46,6 +46,7 @@ public class MainActivity extends AppCompatActivity {
                 })
                 .setOnConsentReady(consent -> {
                     Log.i(TAG, "onConsentReady");
+                    Log.i(TAG, "consentString: " + (consent.consentString != null ? consent.consentString : "<empty>"));
                     for (String vendorId : consent.acceptedVendors) {
                         Log.i(TAG, "The vendor " + vendorId + " was accepted.");
                     }
@@ -82,8 +83,8 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onResume() {
         super.onResume();
-        buildGDPRConsentLib().run(buildNativeMessage());
-        //buildGDPRConsentLib().run() can be called (with no arg) in order to work with the webview based message
+        Log.i(TAG, "init");
+        buildGDPRConsentLib().run();
     }
 
     @Override