+++ cmplibrary/src/main/java/com/sourcepoint/gdpr_cmplibrary/GDPRConsentLib.java
@@ -157,9 +157,13 @@ public class GDPRConsentLib {
 
     }
 
+    public void clearAllData(){
+        storeClient.clearAllData();
+    }
+
     private void setConsentData(String newAuthId){
 
-        if(hasAuthIdChanged(newAuthId)) storeClient.clearAllData();
+        if(didAuthIdChange(newAuthId)) clearAllData();
 
         euConsent = storeClient.getConsentString();
 
@@ -170,7 +174,7 @@ public class GDPRConsentLib {
         storeClient.setAuthId(newAuthId);
     }
 
-    private boolean hasAuthIdChanged(String newAuthId){
+    private boolean didAuthIdChange(String newAuthId){
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
             return !Objects.equals(newAuthId, storeClient.getAuthId());
         }
@@ -185,8 +189,7 @@ public class GDPRConsentLib {
         return new ConsentWebView(activity, defaultMessageTimeOut) {
 
             @Override
-            public void onMessageReady() {
-                Log.d("msgReady", "called");
+            public void onConsentUIReady() {
                 cancelCounter();
                 runOnLiveActivityUIThread(() -> GDPRConsentLib.this.onConsentUIReady.run(this));
             }
@@ -199,18 +202,18 @@ public class GDPRConsentLib {
             @Override
             public void onSavePM(GDPRUserConsent u) {
                 GDPRConsentLib.this.userConsent = u;
-                closeAllView();
-                sendConsent(ActionTypes.PM_COMPLETE);
+                closeAllViews();
+                sendConsent(ActionTypes.PM_COMPLETE, null);
             }
 
             @Override
-            public void onAction(int choiceType) {
-                GDPRConsentLib.this.onAction(choiceType, 0);
+            public void onAction(int choiceType, Integer choiceId) {
+                GDPRConsentLib.this.onAction(choiceType, choiceId);
             }
         };
     }
 
-    public void onAction(int choiceType, int choiceId) {
+    public void onAction(int choiceType, Integer choiceId) {
         try{
             Log.d(TAG, "onAction:  " +  choiceType  + " + choiceType");
             switch (choiceType) {
@@ -218,13 +221,13 @@ public class GDPRConsentLib {
                     onMsgShowOptions();
                     break;
                 case ActionTypes.MSG_ACCEPT:
-                    onMsgAccepted();
+                    onMsgAccepted(choiceId);
                     break;
                 case ActionTypes.MSG_CANCEL:
-                    onMsgCancel();
+                    onMsgCancel(choiceId);
                     break;
                 case ActionTypes.MSG_REJECT:
-                    onMsgRejected();
+                    onMsgRejected(choiceId);
                     break;
                 case ActionTypes.PM_DISMISS:
                     onPmDismiss();
@@ -232,7 +235,6 @@ public class GDPRConsentLib {
                     break;
             }
         } catch (Exception e) {
-            e.printStackTrace();
             GDPRConsentLib.this.onErrorTask(new ConsentLibException(e, "Unexpected error when calling onAction."));
         }
     }
@@ -242,9 +244,9 @@ public class GDPRConsentLib {
         nativeView.setAttributes(new NativeMessageAttrs(msgJson));
     }
 
-    public void onMsgAccepted() {
-        closeAllView();
-        sendConsent(ActionTypes.MSG_ACCEPT);
+    public void onMsgAccepted(Integer choiceId) {
+        closeAllViews();
+        sendConsent(ActionTypes.MSG_ACCEPT, choiceId);
     }
 
     protected  void onPmDismiss(){
@@ -262,29 +264,42 @@ public class GDPRConsentLib {
         return isNative ? nativeView : webView;
     }
 
-    public void onMsgCancel(){
-
+    public void onMsgCancel(Integer choiceId){
         closeCurrentMessageView();
         consentFinished();
     }
 
-    public void onMsgRejected() {
-        closeAllView();
-        sendConsent(ActionTypes.MSG_REJECT);
+    public void onMsgRejected(Integer choiceId) {
+        closeAllViews();
+        sendConsent(ActionTypes.MSG_REJECT, choiceId);
     }
 
     public void onMsgShowOptions(){
-        try{
-            loadPm();
-        } catch(Exception e){
-            onErrorTask(new ConsentLibException(e, "Unexpexted error trying to show PM"));
-        }
+        loadPm();
     }
 
     private void loadPm() {
-        if(webView == null) webView = buildWebView();
-        webView.loadUrl(pmUrl());
+        loadConsentUI(pmUrl());
         isPmOn = true;
+    }
+
+    private void loadConsentUI(String url){
+        mCountDownTimer = getTimer(defaultMessageTimeOut);
+        mCountDownTimer.start();
+        if(webView == null) {
+            webView = buildWebView();
+            webView.loadConsentUIFromUrl(url);
+        } else if(!isNative) {
+            webView.post(new Runnable() {
+                @Override
+                public void run() {
+                    webView.loadConsentUIFromUrl(url);
+                }
+            });
+        } else {
+            showView(webView);
+            cancelCounter();
+        }
     }
 
 
@@ -298,8 +313,6 @@ public class GDPRConsentLib {
     public void run() {
         try {
             onMessageReadyCalled = false;
-            mCountDownTimer = getTimer(defaultMessageTimeOut);
-            mCountDownTimer.start();
             renderMsgAndSaveConsent();
         } catch (Exception e) {
             e.printStackTrace();
@@ -317,7 +330,6 @@ public class GDPRConsentLib {
             isNative = true;
             renderMsgAndSaveConsent();
         } catch (Exception e) {
-            e.printStackTrace();
             onErrorTask(new ConsentLibException(e, "Error trying to load pm URL."));
         }
     }
@@ -330,15 +342,15 @@ public class GDPRConsentLib {
                     JSONObject jsonResult = (JSONObject) result;
                     consentUUID = jsonResult.getString("uuid");
                     metaData = jsonResult.getString("meta");
-                    userConsent = new GDPRUserConsent(jsonResult.getJSONObject("userConsent"));
                     if(jsonResult.has("msgJSON") && !jsonResult.isNull("msgJSON")) {
                         setNativeMessageView(jsonResult.getJSONObject("msgJSON"));
                         showView(nativeView);
                     } else if(jsonResult.has("url") && !jsonResult.isNull("url")){
-                        webView = buildWebView();
-                        webView.loadConsentUIFromUrl(jsonResult.getString("url"));
+                        loadConsentUI(jsonResult.getString("url"));
                     } else {
-                        cancelCounter();
+                        userConsent = new GDPRUserConsent(jsonResult.getJSONObject("userConsent"));
+                        if(euConsent == null) euConsent = userConsent.consentString;
+                        else userConsent.consentString = euConsent;
                         consentFinished();
                     }
                 }
@@ -360,7 +372,7 @@ public class GDPRConsentLib {
         runOnLiveActivityUIThread(() -> GDPRConsentLib.this.onConsentUIReady.run(view));
     }
 
-    public void closeAllView(){
+    public void closeAllViews(){
         if(isNative){
             closeView(nativeView);
             if(isPmOn) closeView(webView);
@@ -377,30 +389,33 @@ public class GDPRConsentLib {
         runOnLiveActivityUIThread(() -> GDPRConsentLib.this.onConsentUIFinished.run(v));
     }
 
-    private JSONObject paramsToSendConsent(int actionType) throws ConsentLibException {
+    private JSONObject paramsToSendConsent(int actionType, Integer choiceId) throws ConsentLibException {
         try{
 
             Log.i("GDPR_UUID", "From sendConsentBody: " + consentUUID);
 
             JSONObject params = new JSONObject();
 
+            //TODO: get rid of jsonConsents, decouple the pm consents json from the userConsents obj
             params.put("consents", userConsent != null ? userConsent.jsonConsents : null);
             params.put("accountId", accountId);
             params.put("propertyId", propertyId);
+            params.put("propertyHref", "https://" + property);
             params.put("privacyManagerId", pmId);
             params.put("uuid", consentUUID);
             params.put("meta", metaData);
             params.put("actionType", actionType);
-            params.put("requestFromPM", true);
+            params.put("requestFromPM", choiceId == null);
+            params.put("choiceId", choiceId != null ? Integer.toString(choiceId) : null);
             return params;
         } catch(JSONException e){
             throw new ConsentLibException(e, "Error trying to build body to send consents.");
         }
     }
 
-    private void sendConsent(int actionType) {
+    private void sendConsent(int actionType, Integer choiceId) {
         try {
-            sourcePoint.sendConsent(paramsToSendConsent(actionType), new OnLoadComplete() {
+            sourcePoint.sendConsent(paramsToSendConsent(actionType, choiceId), new OnLoadComplete() {
                 @Override
                 public void onSuccess(Object result) {
                     try{